requirements.txt
requests>=2.32,<3
python-dotenv>=1.0,<2
pandas
pydantic
pydantic-ai
pytrends
duckduckgo-search
